requirements.txt
ydb
python-telegram-bot==20.0a0
beautifulsoup4
